+++ xdsdk_docs/Unity/CHANGELOG.md
@@ -1,6 +1,9 @@
 #Change Log
 
 心动SDK_Unity 的 CHANGELOG 记录每次版本更新日志  面向开发
+###Unity 1.1.9 - 2018-07-30（iOS-3.2.20&Android-3.2.10)
+- 更新Android3.2.10
+
 ###Unity 1.1.8 - 2018-07-24（iOS-3.2.20&Android-3.2.9)
 - 更新taptapSDK
 - iPad未安装微信显示微信登录入口，点击跳转扫码登录